% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_summaries.R
\name{asb.sp.summary}
\alias{asb.sp.summary}
\title{Summarize assemblage*species dataframe}
\usage{
asb.sp.summary(asb_sp_w)
}
\arguments{
\item{asb_sp_w}{a \strong{matrix} showing assemblages (rows) composition in
species (columns).}
}
\value{
a species occurrences data.frame in each assemblage
\code{asb_sp_w_occ}. Use the Data Viewer to see it ; a vector gathering
species biomass/abundance per species \code{tot_ab_all_sp} ; a vector
gathering total abundance/biomass per assemblage \code{tot_ab_all_asb} ; a
vector gathering species richness per assemblage \code{sp_richn_all_asb} ;
a list gathering the names of species of each assemblage \code{sp_nm_asb}.
}
\description{
Computes summary helping you to picture assemblages. For this function to
work, there must be no NA in your \code{assemblage_species} dataframe.
}
\examples{
load(system.file("extdata", "asb_sp_w_fruits", package = "mFD"))
asb_sp_w <- as.matrix(asb_sp_w)
mFD::asb.sp.summary(asb_sp_w)

}
