% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asb_sp_w_fruits.R
\docType{data}
\name{asb_sp_w_fruits}
\alias{asb_sp_w_fruits}
\title{Species assemblages matrix for fruits baskets example}
\format{
A matrix with 10 rows and 25 columns
\describe{
\item{asb}{10 fruits baskets}
\item{sp}{25 fruits species}
}
}
\usage{
asb_sp_w_fruits
}
\description{
A matrix gathering 25 fruits species into 10 fruits baskets.
Each number in the matrix is fruits abundance. This matrix is used in most
mFD tutorials and diverse mFD functions examples.
}
\keyword{datasets}
