% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baskets_fruits_weights.R
\docType{data}
\name{baskets_fruits_weights}
\alias{baskets_fruits_weights}
\title{Baskets Composition in Fruits Species}
\format{
A matrix of integers with 10 rows (baskets) and 25 columns (species).
}
\usage{
baskets_fruits_weights
}
\description{
This dataset represents the abundance of 25 fruits species in 10 baskets.
}
\examples{
# Load Assemblages x Species Matrix
data("baskets_fruits_weights", package = "mFD")
baskets_fruits_weights[1:5, 1:5]

# Summarize Assemblages Data
mFD::asb.sp.summary(asb_sp_w = baskets_fruits_weights)
}
\seealso{
\code{fruits_traits}, \code{fruits_traits_cat}
}
\keyword{datasets}
