% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_used_for_ind_comp.R
\name{dist.nearneighb}
\alias{dist.nearneighb}
\title{Compute distance of a given point to its nearest neighbor in the functional
space and the identity of the nearest neighbor}
\usage{
dist.nearneighb(sp_faxes_coord, ref_sp)
}
\arguments{
\item{sp_faxes_coord}{a matrix of species coordinates in a chosen
functional space. Species coordinates have been retrieved thanks to
\code{\link{tr.cont.fspace}} or \code{\link{quality.fspaces}}.}

\item{ref_sp}{a character string referring to the name of the reference
species}
}
\value{
a list containing the nearest neighbor identity \code{nn_id} and a
list of the distance of the reference point to its nearest neighbor
\code{nn_ref_sp_dist}.
}
\description{
This function is used in functional indices computation.
}
\author{
Camille Magneville and Sébastien Villéger
}
