% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_indiv_ind_computation.R
\name{feve.computation}
\alias{feve.computation}
\title{Compute functional evenness (FEve)}
\usage{
feve.computation(
  asb_sp_relatw_k,
  sp_faxes_coord_k,
  k,
  check_input = check_input
)
}
\arguments{
\item{asb_sp_relatw_k}{a \strong{matrix} containing species relative weight (columns)
for a given assemblage.}

\item{sp_faxes_coord_k}{a matrix of species coordinates present in a
given assemblage in a chosen functional space with only needed axes.
Species coordinates have been retrieved thanks to \code{\link{tr.cont.fspace}} or
\code{\link{quality.fspaces}} and filtered thanks to \code{\link{sp.filter}}.}

\item{k}{a \strong{character string} referring to the assemblage studied.}

\item{check_input}{a \strong{logical value} allowing to test or not the inputs.
Possible error messages will thus may be more understandable for the user
than R error messages. Species coordinates matrix and species*weight
dataframe must not contain NA, their rownames must be filled, they must
have similar names values, and the number of species in the assemblage must
be higher than three to compute feve. Default: check_input = FALSE.}
}
\value{
a matrix containing functional evenness for a given assemblage.
}
\description{
This function computes the regularity of distribution of species weights in
the functional space.
}
\author{
Camille Magneville and Sébastien Villéger
}
