% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_indiv_ind_computation.R
\name{fide.computation}
\alias{fide.computation}
\title{Compute functional identity}
\usage{
fide.computation(
  asb_sp_relatw_k,
  sp_faxes_coord_k,
  k,
  check_input = check_input
)
}
\arguments{
\item{asb_sp_relatw_k}{a \strong{matrix} containing species relative weight (columns)
for a given assemblage.}

\item{sp_faxes_coord_k}{a \strong{matrix} of species coordinates present in a given
assemblage in a chosen functional space with only needed axes. Species
coordinates have been retrieved thanks to \code{fspace.conttr} or
\code{\link{quality.fspaces}} and filtered thanks to \code{\link{sp.filter}}.}

\item{k}{a \strong{character string} referring to the assemblage studied.}

\item{check_input}{a \strong{logical value} allowing to test or not the inputs.
Possible error messages will thus may be more understandable for the user
than R error messages. Species coordinates matrix and species*weight
dataframe must not contain NA, their rownames must be filled and they must
have similar names values. Default: check_input = FALSE.}
}
\value{
a matrix containing functional identity values for a given assemblage
along the dimensions (columns). Number of dimensions is fixed to the number
of dimensions in \code{sp_faxes_coord} data.frame.
}
\description{
This function computes the weighted average position along each axis. FIde is
computed using relative weight so that it is not affected by unit (e.g. g or
kg for biomass). In the special case where 'weight' is filled with only 0/1
(absence/presence), then FIde will be computed assuming that all species have
the same weight. The results of this function are used in FSpe, FOri and FNND
computation.
}
\author{
Camille Magneville and Sébastien Villéger
}
