% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_indiv_ind_computation.R
\name{fnnd.computation}
\alias{fnnd.computation}
\title{Compute functional mean nearest neighbor distance (FNND)}
\usage{
fnnd.computation(
  asb_sp_relatw_k,
  sp_faxes_coord_k,
  k,
  check_input = check_input
)
}
\arguments{
\item{asb_sp_relatw_k}{a \strong{matrix} containing species relative weight (columns)
for a given assemblage.}

\item{sp_faxes_coord_k}{a \strong{matrix} of species coordinates present in a given
assemblage in a chosen functional space with only needed axes. Species
coordinates have been retrieved thanks to \code{\link{tr.cont.fspace}} or
\code{\link{quality.fspaces}} and filtered thanks to \code{\link{sp.filter}}.}

\item{k}{a \strong{character string} referring to the assemblage studied.}

\item{check_input}{a \strong{logical value} allowing to test or not the inputs.
Possible error messages will thus may be more understandable for the user
than R error messages. Species coordinates matrix and species*weight
dataframe must not contain NA, their rownames must be filled and they must
have similar names values. Default: check_input = FALSE.}
}
\value{
a matrix containing functional mean nearest neighbor distance for a
given assemblage.
}
\description{
This function computes the weighted mean distance to nearest neighbor. It
uses /code{dist.nearneighb} function that computes distance to the nearest
neighbor for each species.FNND value can be scaled by the maximum value
possible given species pool (i.e. the most distant species pair have total
weight) to standardize values.
}
\author{
Camille Magneville and Sébastien Villéger
}
