% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multidim_layers.R
\name{fnnd.plot}
\alias{fnnd.plot}
\title{Plot FNND index}
\usage{
fnnd.plot(
  ggplot_bg,
  asb_sp_coord2D,
  asb_sp_relatw,
  asb_nn_asb,
  plot_sp = TRUE,
  shape_sp,
  color_sp,
  fill_sp,
  color_segment,
  width_segment,
  linetype_segment
)
}
\arguments{
\item{ggplot_bg}{a ggplot object of the plot background retrieved through
the \code{\link{background.plot}} function.}

\item{asb_sp_coord2D}{a list of matrix (ncol = 2) with coordinates of
species present in each assemblage for a given pair of functional axes}

\item{asb_sp_relatw}{a list of matrix gathering species relative weight in
each assemblage. It can be retrieved through the
\code{\link{asb.sp.summary}} or \code{\link{alpha.fd.multidim}}}

\item{asb_nn_asb}{a list gathering for each species of a studied assemblage
its nearest neighbour(s) in the assemblage. If several assemblages it
should  be a vector with names as in \code{asb_sp_coord2D}.}

\item{plot_sp}{a logical value indicating whether species of each assemblage
should be plotted or not. Default: \code{plot_sp = TRUE}}

\item{shape_sp}{a numeric value referring to the shape used to plot species
belonging to the studied assemblage. Default: \code{shape_sp = 16}
(filled circle). If several assemblages it should be a vector with
names as in \code{asb_sp_coord2D}.}

\item{color_sp}{a R color name or an hexadecimal code referring to the color
of species from the studied assemblage. Default: \code{color_sp = "#0072B2"}.
If several assemblages it should be a vector with names as in
\code{asb_sp_coord2D}.}

\item{fill_sp}{a R color name or an hexadecimal code referring to the colour
to fill species symbol (if \code{shape_sp} > 20). Default:
\code{fill_sp = '#0072B2'}. If several assemblages it should  be a vector with
names as in \code{asb_sp_coord2D}.}

\item{color_segment}{a R color name or an hexadecimal code referring to the
color of the segment linking nearest neighbors in the studied assemblage.
If several assemblages it should be a vector with names as in
\code{asb_sp_coord2D}.}

\item{width_segment}{a numeric value referring to the size of the segment
linking nearest neighbors in the studied assemblage. Default:
\code{width_segment = 1}. If several assemblages it should be a vector with
names as in \code{asb_sp_coord2D}.}

\item{linetype_segment}{a character string referring to the linetype used to
link nearest neighbors in the studied assemblages.
Default: \verb{linetype_segment = "dashed}.  If several assemblages it should be a
vector with names as in \code{asb_sp_coord2D}.}
}
\value{
a ggplot object with FNND index
}
\description{
This function plots FNND index for a given pair of functional axes and for
one or several assemblages
}
\examples{
\dontrun{
}
}
\author{
Camille Magneville and Sébastien Villéger
}
