% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_indiv_ind_computation.R
\name{fori.computation}
\alias{fori.computation}
\title{Compute functional originality}
\usage{
fori.computation(
  dist_nn_global_pool,
  asb_sp_relatw_k,
  k,
  check.input = check.input
)
}
\arguments{
\item{dist_nn_global_pool}{a \bold{vector} containing the minimal distance to the
nearest neighbor for each species of the global pool of species}

\item{asb_sp_relatw_k}{a \bold{matrix} containing species relative weight
(columns) for a given assemblage.}

\item{k}{a \bold{character string} referring to the assemblage studied.}

\item{check.input}{a \bold{logical value} allowing to test or not the inputs.
Possible error messages will thus may be more understandable for the user
than R error messages. Species*weight dataframe must not contain NA and its
rownames must be filled. Default: check.input = FALSE.}
}
\value{
a matrix containing functional originality for a given assemblage.
}
\description{
This function computes the weighted mean distance to nearest species from the
species pool. FOri value can be scaled by the maximum distance to the nearest
neighbour possible in the global species pool (i.e. an assemblage hosting
only the most original species).
}
