% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_indiv_ind_computation.R
\name{fric.computation}
\alias{fric.computation}
\title{Compute functional richness}
\usage{
fric.computation(sp_faxes_coord_k, k, check.input = check.input)
}
\arguments{
\item{sp_faxes_coord_k}{a \bold{matrix} of species coordinates present in a given
assemblage in a chosen functional space with only needed axes. Species
coordinates have been retrieved thanks to \code{fspace.conttr} or
\code{qual.funct.space} and filtered thanks to \code{sp.filter}.}

\item{k}{a \bold{character string} referring to the assemblage studied.}

\item{check.input}{a \bold{logical value} allowing to test or not the inputs.
Possible error messages will thus may be more understandable for the user
than R error messages. Species coordinates matrix must not contain NA, its
rownames must be filled and the number of species should strictly be higher
than the number of axes to compute the convex hull. Default: check.input =
FALSE.}
}
\value{
a list containing: \strong{$fric} a vector with fric value for a
given assemblage and \strong{vertices_nm} a vector containing names of the
species being vertices (species are ordered as in row names of input)
}
\description{
This function computes the volume of functional space filled by species
present in a given assemblage.
}
\note{
Computation with qconvex algorithm is led using option "Tv" so result
are verified for structure, convexity, and point inclusion. FRic value is
based on axes units. See \code{\link{alpha.fd.multidim}} for option to scale
values using volume by species pool.
}
