% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_indiv_ind_computation.R
\name{fspe.computation}
\alias{fspe.computation}
\title{Compute functional specialization}
\usage{
fspe.computation(
  asb_sp_relatw_k,
  special_sp_global_pool,
  k,
  check_input = check_input
)
}
\arguments{
\item{asb_sp_relatw_k}{a \strong{matrix} containing species relative weight
(columns) for a given assemblage.}

\item{special_sp_global_pool}{a \strong{vector} containing the distance to the
centroid of the global pool of species for each species of the global pool
(it is called specialization).}

\item{k}{a \strong{character string} referring to the assemblage studied.}

\item{check_input}{a \strong{logical value} allowing to test or not the inputs.
Possible error messages will thus may be more understandable for the user
than R error messages. Species*weight dataframe must not contain NA and its
rownames must be filled. Default: check_input = FALSE.}
}
\value{
a matrix containing functional specialization for a given assemblage.
}
\description{
This function computes the weighted mean distance to the centroid of the
global species pool. It computes the average position of all the species
present given assemblage. FSpe value can be scaled by the maximum distance to
the global pool centroid (i.e. an assemblage hosting only the most
specialized species).
}
\author{
Camille Magneville and Sébastien Villéger
}
