% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse_ind_computation.R
\name{get.indicator}
\alias{get.indicator}
\title{Compute the nearest neighbors for all considered species for FUSE computation}
\usage{
get.indicator(sp_dist, nb_NN)
}
\arguments{
\item{sp_dist}{a matrix object representing the distance in an euclidean
space between species based on species traits}

\item{nb_NN}{a numerical value giving the number of nearest neighbor to
consider}
}
\description{
Compute the nearest neighbors for all considered species for FUSE computation
}
