% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFD_utility.R
\name{get_cs_index_vector}
\alias{get_cs_index_vector}
\title{Map credible set labels to a full SNP vector}
\usage{
get_cs_index_vector(cs_res, n_snps, renumber = TRUE)
}
\arguments{
\item{cs_res}{Result object containing \code{cs} and optionally \code{cs_index}.}

\item{n_snps}{Total number of SNPs.}

\item{renumber}{If \code{TRUE}, renumber sets sequentially.}
}
\value{
Integer vector of length \code{n_snps}.
}
\description{
Map credible set labels to a full SNP vector
}
