% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse_ind_computation.R
\name{get_dist_func}
\alias{get_dist_func}
\title{Calculate the nearest neighbors for a given species}
\usage{
get_dist_func(nb_NN, data)
}
\arguments{
\item{nb_NN}{a numerical value giving the number of nearest neighbor to consider}

\item{data}{a dataframe considering all the distance between the considered
species and all of these neighbors}
}
\description{
Calculate the nearest neighbors for a given species
}
