% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFD_utility.R
\name{in_CS}
\alias{in_CS}
\title{Compute credible set membership}
\usage{
in_CS(res, coverage = 0.95)
}
\arguments{
\item{res}{Numeric matrix of posterior probabilities.}

\item{coverage}{Credible set coverage threshold.}
}
\value{
A matrix of 0/1 indicators.
}
\description{
Applies a credible set membership rule row-wise to a matrix of posterior
probabilities.
}
