% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFD_utility.R
\docType{package}
\name{MFD-package}
\alias{MFD}
\alias{MFD-package}
\title{MFD: Decision-guided multi-ancestry fine-mapping}
\description{
Utilities for decision-guided multi-ancestry fine-mapping using
SuSiE post-hoc and MESuSiE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gchen4422/MFD}
  \item Report bugs at \url{https://github.com/gchen4422/MFD/issues}
}

}
\author{
\strong{Maintainer}: Guangxin Chen \email{chen4422@purdue.edu}

}
\keyword{internal}
