% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multidim_layers.R
\name{panels.to.patchwork}
\alias{panels.to.patchwork}
\title{Plot individual plots along a pair of functional axes into a unique graph}
\usage{
panels.to.patchwork(panels, plot_caption)
}
\arguments{
\item{panels}{a list of ggplot objects illustrating an index on two given
functional axes. There must be either one, three or six ggplot objects given
the number of studied functional axes.}

\item{plot_caption}{a ggplot object illustrating the caption of the
final patchwork plot.}
}
\value{
a unique ggplot object gathering functional ppanels and caption
}
\description{
This function gathers panels into a unique \code{patchwork} graph
with caption
}
\author{
Camille Magneville and Sébastien Villéger
}
