% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_fspaces.R
\name{quality.fspaces.plot}
\alias{quality.fspaces.plot}
\title{Plot functional space quality with a chosen quality metric}
\usage{
quality.fspaces.plot(
  fspaces_quality,
  quality_metric,
  fspaces_plot,
  name_file = NULL,
  range_dist = NULL,
  range_dev = NULL,
  range_qdev = NULL,
  gradient_deviation = c(neg = "darkblue", nul = "grey80", pos = "darkred"),
  gradient_deviation_quality = c(low = "yellow", high = "red"),
  x_lab = "Trait-based distance"
)
}
\arguments{
\item{fspaces_quality}{output from the \code{\link{quality.fspaces}} function,
that is a list with all data needed to illustrate quality of functional
spaces based on deviation between species trait-based distance and distance
in functional spaces built using PCoA (and dendrogram).}

\item{quality_metric}{a \strong{character string} with the name of the quality metric
to illustrate. Should be one of the column names of
\code{fspaces_quality$quality_fspaces}. See help of \code{\link{quality.fspaces}}
for meaning of these names regarding type of deviation and scaling of
distance in functional space. Default: 'mad' (Mean absolute deviation).}

\item{fspaces_plot}{a \strong{vector} with names of functional spaces to consider.
Should be a subset of the row names of
\code{fspaces_quality$quality_fspaces}. Maximum of 10 spaces allowed to keep decent
plot size.}

\item{name_file}{a \strong{character string} with name of file to save the figure
(without extension). Default: NULL which means plot is displayed.}

\item{range_dist}{a \strong{vector} with minimum and maximum values to display for
species pairwise distances (x-axis fo all panels and y-axes of top panel).
Default: NULL, which means range is 0 to maximum distance among all the
functional spaces to plot.}

\item{range_dev}{a \strong{vector} with minimum and maximum values to display for
deviation to trait-based distance (y-axis of middle panel). Default: NULL,
which means range is set to range of deviation among all the functional
spaces to plot.}

\item{range_qdev}{a \strong{vector} with minimum and maximum values to display for
deviation to trait-based distance (y-axis of bottom panel). Default:NULL,
which means range is from 0 to the maximum of (transformed) deviation among
all the functional spaces to plot.}

\item{gradient_deviation}{a \strong{vector} of 3 colors for illustrating raw deviation
with \code{\link[ggplot2]{scale_colour_gradient2}}. First value ('neg') is for the
lowest negative deviation, second value ('nul')is for null deviation and
third value ('pos') is for the highest positive deviation. Default gradient
is from darkblue to grey to red.}

\item{gradient_deviation_quality}{2 colors (named 'low' and 'high') for
illustrating transformed deviation used to compute quality metric with
\code{\link[ggplot2]{scale_colour_gradient2}} (default gradient is from yellow to red).}

\item{x_lab}{a \strong{character string} with title to display below X axis. Default
is 'Trait-based distance'.}
}
\value{
a png file (resolution 300dpi) saved in the current working
directory. Quality of each functional space is illustrated with three
panels : - top row shows trait-based distance between species vs.
space-based distance. - middle row shows trait-based distance vs. deviation
between space-based and trait-based distances - bottom row shows
trait-based distance between species vs. transformed deviation used to
compute the quality metric All plots have same X axis. All plots on a given
row have the same Y axis and color palette. Type of distance in functional
space (Euclidean in PCoA, Cophenetic on tree) are abbreviated, as well as
type of transformation of distance (scaling) and of deviation (Absolute or
Squared)
}
\description{
Plot functional space quality with a chosen quality metric
}
\examples{
library(patchwork)
load(system.file("extdata", "sp_tr_fruits_df", package = "mFD"))
sp_tr <- sp_tr[, -c(6:8)]
load(system.file("extdata", "sp_tr_cat_fruits_df", package = "mFD"))
sp_tr_cat <- sp_tr_cat[-c(6:8), ]
sp_trdist <- mFD::funct.dist(sp_tr, sp_tr_cat, dist_metric = "classical_gower")
fspaces_quality <- mFD::quality.fspaces(sp_trdist, maxdim_pcoa = 10,
 deviation_weighting = "absolute", fdist_scaling = FALSE, fdendro = "average")
mFD::quality.fspaces.plot(fspaces_quality, quality_metric = "mad",
 fspaces_plot = c("tree_average", "pcoa_2d", "pcoa_3d", "pcoa_4d", "pcoa_5d"),
 name_file = NULL, range_dist = NULL, range_dev = NULL, range_qdev = NULL,
 gradient_deviation  = c(neg = "darkblue", nul = "grey80", pos = "darkred"),
 gradient_deviation_quality = c(low ="yellow", high = "red"),
 x_lab = "Trait-based distance")

}
