% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multidim_layers.R
\name{sp.plot}
\alias{sp.plot}
\title{Plot species}
\usage{
sp.plot(
  ggplot_bg,
  asb_sp_coord2D,
  asb_vertices_nD = NULL,
  asb_sp_relatw = NULL,
  size_sp,
  shape_sp,
  color_sp,
  fill_sp,
  size_vert,
  shape_vert,
  color_vert,
  fill_vert,
  limits_relatw = c(0, 1),
  range_size_relatw = c(1, 10)
)
}
\arguments{
\item{ggplot_bg}{a ggplot object of the plot background retrieved through
the \code{\link{background.plot}} function.}

\item{asb_sp_coord2D}{a list of matrix (ncol = 2) with coordinates of
species present in each assemblage for a given pair of axes.}

\item{asb_vertices_nD}{a list (with names as in asb_sp_coord2D) of vectors
with names of species being vertices in n dimensions.}

\item{asb_sp_relatw}{a list of vector gathering species relative weight in
each assemblage. It can be retrieved through the
\code{\link{alpha.fd.multidim}}.}

\item{size_sp}{a numeric value referring to the size of species belonging to
the plotted assemblage. Default: \code{size_sp = 1}.}

\item{shape_sp}{a numeric value referring to the shape of species
belonging to the plotted assemblage.
Default: \code{shape_sp = 16}.}

\item{color_sp}{a R color name or an hexadecimal code referring to the color
of species from the studied assemblage.  This color is also used for FRic
convex hull color.}

\item{fill_sp}{a R color name or an hexadecimal code referring to the colour
to fill species symbol (if \code{shape_sp} > 20) and the assemblage convex
hull.}

\item{size_vert}{a numeric value referring to the size of symbol for vertices
Default: \code{size_vert = 1}.}

\item{shape_vert}{a numeric value referring to the shape of symbol for
vertices.}

\item{color_vert}{a R color name or an hexadecimal code referring to the
color of vertices if plotted. If color_vert = NA, vertices are not plotted
(for shapes only defined by color, ie shape inferior to 20. Otherwise fill
must also be set to NA).}

\item{fill_vert}{a character value referring to the color for filling symbol
for vertices (if \code{shape_vert} >20). If \code{fill = NA} and \code{color = NA},
vertices are not plotted (if \code{shape_vert} superior to 20. Otherwise
\code{color_vert = NULL} is enough).}

\item{limits_relatw}{a vector of two numbers giving the limits to set for
the scale of species relative weights.}

\item{range_size_relatw}{a vector of two numbers specifying the minimum and
the maximum size of the plotting symbol for relative weights.}
}
\value{
a ggplot object with species plotted on the background plot
}
\description{
Plot species for one to n assemblages
}
\note{
if \code{asb_vertices_nD = NULL}, all arguments for vertices are
ignored and if \code{asb_sp_relatw != NULL}, \code{size_sp} and
\code{size_vert} are ignored. And if several assemblages are to be
represented, aesthetics inputs should be formatted as c(pool = ..., asb1 =
..., asb2 = ...).
}
\author{
Camille Magneville and Sébastien Villéger
}
