% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_summaries.R
\name{sp.tr.summary}
\alias{sp.tr.summary}
\title{Summarize species*traits dataframe}
\usage{
sp.tr.summary(tr_cat, sp_tr)
}
\arguments{
\item{tr_cat}{a \bold{dataframe} containing three columns for each trait (rows):
\itemize{ \item \strong{trait_name}: names of all traits as in \code{sp_tr}
data.frame \item \strong{trait_type}: category codes for each trait as
followed: \emph{N} for Nominal traits (factor variable), \emph{O} for
Ordinal traits (ordered variable), \emph{C} for Circular traits (integer
values), \emph{Q} for quantitative traits (numeric values) and
\emph{F} for fuzzy traits (described with several values defined with
several column). } An option is to add a fourth column with a numeric
vector of length n (traits number) to specify a weight for each trait.}

\item{sp_tr}{a \bold{dataframe} of traits values (columns) for each species (rows)}
}
\value{
a table summarizing for each trait the number of species per modality
for non-continuous trait and min/max/mean/median/quartiles for continuous
traits \code{tr_summary_list} ; a list containing traits type
\code{tr_types} ; a list containing modalities for each non continuous
trait \code{non_conttr_mod_list}
}
\description{
Computes summary data helping to choose the type of analysis you can do with
your data. For this function to work, there must be no NA in your
\code{sp_tr} dataframe.
}
\examples{
 load(system.file("extdata", "sp_tr_fruits_df", package = "mFD"))
 load(system.file("extdata", "sp_tr_cat_fruits_df", package = "mFD"))
 mFD::sp.tr.summary(sp_tr_cat, sp_tr)

}
