% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_tr.R
\docType{data}
\name{sp_tr}
\alias{sp_tr}
\title{Species traits dataframe for the fruits baskets example}
\format{
dataframe with 25 rows and 8 columns
\describe{
\item{sp}{25 fruits species}
\item{tr}{6 traits}
}
}
\usage{
sp_tr
}
\description{
A dataframe gathering 25 fruits species (rows) and values for 6
traits (columns) used in all mFD tutorials and divers functions examples
}
\keyword{datasets}
