% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_tr_cat_fruits.R
\docType{data}
\name{sp_tr_cat_fruits}
\alias{sp_tr_cat_fruits}
\title{Dataframe of traits category for fruits example}
\format{
A dataframe with 8 rows and 3 columns
\describe{
\item{tr}{6 traits of different types}
\item{tr_cat}{traits names, traits types and names for fuzzy coded traits}
}
}
\usage{
sp_tr_cat_fruits
}
\description{
A dataframe summarizing traits types for each trait of the
fruits baskets dataset. First column gathers traits names, second columns
gathers traits types, third column summarizes to which fuzzy trait belongs
each column that refers to a fuzzy trait.
}
\keyword{datasets}
