% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_tr_fruits.R
\docType{data}
\name{sp_tr_fruits}
\alias{sp_tr_fruits}
\title{Species traits dataframe linking fruits and traits for mFD use}
\format{
A dataframe with 25 rows and 8 columns
\describe{
\item{sp}{25 fruits species}
\item{tr}{6 traits of different types}
}
}
\usage{
sp_tr_fruits
}
\description{
A dataframe linking 25 fruits species (rows) and their traits
values for 6 functional traits (columns). It is used in all mFD tutorials
and diverse mFD functions examples
}
\keyword{datasets}
