% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_tr_fspace.R
\name{tr.cont.fspace}
\alias{tr.cont.fspace}
\title{Build functional space}
\usage{
tr.cont.fspace(
  sp_tr,
  pca = TRUE,
  nb_dim = 7,
  scaling = "scale_center",
  compute_corr = "pearson"
)
}
\arguments{
\item{sp_tr}{a \strong{dataframe} of raw/standardized traits values for each species.
Here traits must be \strong{continuous}}

\item{pca}{a \strong{logical value}, \code{TRUE} to compute PCA analysis, \code{FALSE}
to compute functional space with one trait for each dimension. Default:
TRUE}

\item{nb_dim}{a \strong{numerical value} referring to the maximum number of dimensions
for multidimensional functional spaces. Final number of dimensions depends
on the number of positive eigenvalues obtained with PCA. High value for
nb_dim can increase computation time. Default: nbdim=7.}

\item{scaling}{a \strong{string value} to compute or not scaling of traits
using the \code{\link{tr.cont.scale}} function. Possible options are
standardizing by the range \code{range}, center standardization
\code{center}, scale standardization \code{scale}, scale center
standardization \code{scale_center} or no scaling \code{no_scale}. Default : scale =}

\item{compute_corr}{a \strong{string value} to compute Pearson correlation
coefficients between traits \code{"pearson"}. You can choose not to compute
correlation coefficient by \code{"none"}}
}
\value{
a dataframe containing species coordinates on each functional axe
\code{sp_faxes_coord}; species distance matrix \code{sp_dist_init};
correlation coefficient between traits if asked.
}
\description{
Compute functional space based on continuous standardized traits or
continuous raw traits matrix. User can either choose to compute functional
space based on PCA analysis or using one trait for one functional axe. For
PCA analysis, center and scale arguments are considered FALSE: if you want to
center, scale or standardize by any mean your data, please use
\code{scale.conttr} function. Option makes it possible to compute correlation
between traits.
}
\examples{
load(system.file("extdata", "sp_tr_cestes_df", package = "mFD"))
mFD::tr.cont.fspace(sp_tr, pca = TRUE, nb_dim = 7, scaling = "scale_center",
 compute_corr = "pearson")

}
