% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_cat_fruits.R
\docType{data}
\name{tr_cat_fruits}
\alias{tr_cat_fruits}
\title{Dataframe gathering traits  for use in mFD example and tutorials}
\format{
A dataframe with 8 rows and 3 columns
\describe{
\item{tr}{8 traits types}
\item{cat}{first column: traits names, second column: traits types, third column: fuzzy traits names}
}
}
\usage{
tr_cat_fruits
}
\description{
A dataframe summarizing traits names with traits names on the first column,
traits_types on the second column and a third column summarizing name of
fuzzy-coded trait to which 'sub-trait' belongs (if trait is not fuzzy,
ignored so could be trait name or NA).
}
\keyword{datasets}
