% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asb_sp_w_fruits.R
\docType{data}
\name{asb_sp_w_fruits}
\alias{asb_sp_w_fruits}
\title{Matrix gathering fruits species into fruits baskets for use in mFD}
\format{
A matrix with 10 rows and 25 columns
\describe{
\item{asb}{10 fruits baskets}
\item{sp}{25 fruits species}
}
}
\usage{
asb_sp_w_fruits
}
\description{
A matrix linking 25 fruits species (columns) and the fruits
basket(s) they belong to (rows). Values represent fruits abundances in each
basket.
}
\keyword{datasets}
