% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asb_sp_w_fruits.R
\docType{data}
\name{asb_sp_w_fruits}
\alias{asb_sp_w_fruits}
\title{Baskets Composition in Fruits Species}
\format{
A matrix of integers with 10 rows (baskets) and 25 columns (species).
}
\usage{
asb_sp_w_fruits
}
\description{
This dataset represents the abundance of 25 fruits species in 10 baskets.
}
\examples{
# Load Assemblages x Species Matrix
data("asb_sp_w_fruits", package = "mFD")
asb_sp_w_fruits[1:5, 1:5]

# Summarize Assemblages Data
mFD::asb.sp.summary(asb_sp_w = asb_sp_w_fruits)
}
\seealso{
\code{sp_tr_fruits}, \code{tr_cat_fruits}
}
\keyword{datasets}
