% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_fhill_ind_computation.R
\name{beta.fd.hill}
\alias{beta.fd.hill}
\title{Compute functional beta-diversity indices based on Hill numbers
applied to distance between species following the framework from Chao et al.
2019, Ecological Monographs (89:e01343), DOI: 10.1002/ecm.1343)}
\usage{
beta.fd.hill(
  asb_sp_w,
  sp_dist,
  q = c(0, 1, 2),
  tau = "mean",
  beta_type = "Jaccard",
  check.input = TRUE,
  store.details = TRUE
)
}
\arguments{
\item{asb_sp_w}{a \strong{matrix} with weight of species (columns) in a set of
assemblages (rows). Rows and columns should have names. NA are not allowed.}

\item{sp_dist}{a \strong{matrix or dist object} with distance between species. Species
names should be provided and match those in 'asb_sp_w'. NA are not allowed.}

\item{q}{a \strong{vector} containing values referring to the order of diversity to use}

\item{tau}{a \strong{character string} with name of function to apply to distance
matrix (i.e. among all pairs of species) to get the threshold used to define
'functionally indistinct set of species'. Could be qet to 'mean' (default),
'min' or 'max'.}

\item{beta_type}{a \strong{character string} with name of framework used for computing
beta-diversity, either 'Jaccard' (default) or 'Sorensen'.}

\item{check.input}{a \strong{logical value} defining whether inputs are checked before
computation of indices. Possible error messages will thus may be more
understandable for the user than R error messages. Default: check.input =
TRUE.}

\item{store.details}{a \strong{logical value} indicating whether the user want to store
values used for computing indices (see list below)}
}
\value{
a list with: \itemize{

\item \emph{asb_FDbeta} a dataframe with a row for each pair of assemblages
(names in 2 first columns, as in \strong{asb_sp_w}) and beta-diversity
for each value of q in other column(s)

\item if \strong{store.details} turned to TRUE a list \emph{details} with
\itemize{
\item \emph{asb_FDalpha} a dataframe with mean alpha diversity of each pair
of assemblages (rows) and values of q (columns)
\item \emph{asb_FDgamma} a dataframe with gamma diversity of each pair
of assemblages (rows) and values of q (columns)
}
}
}
\description{
Compute functional beta-diversity indices based on Hill numbers
applied to distance between species following the framework from Chao et al.
2019, Ecological Monographs (89:e01343), DOI: 10.1002/ecm.1343)
}
\note{
when q=1 Jaccard-like and Sorensen-like beta-diversity are identical.
FD computed with tau='min' is equivalent to Hill number taxonomic beta
diversity.
}
\examples{
load(system.file("extdata", "sp_tr_fruits_df", package = "mFD"))
sp_tr <- sp_tr[, -c(6:8)]
load(system.file("extdata", "asb_sp_w_fruits", package = "mFD"))
asb_sp_w <- as.matrix(asb_sp_w)
sp_dist <- cluster::daisy(sp_tr, metric = "gower")
 beta.fd.hill(asb_sp_w, sp_dist, q = c(0,1,2), tau = "mean",
 beta_type = "Jaccard", check.input = TRUE, store.details = TRUE)

}
