% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_ind_computation.R
\name{beta.fd.multidim}
\alias{beta.fd.multidim}
\title{Computes functional beta-diversity indices for pairs of assemblages in a
multidimensional space.}
\usage{
beta.fd.multidim(
  sp_faxes_coord,
  asb_sp_occ,
  check.input = TRUE,
  beta.family = "Jaccard",
  store_details = TRUE,
  betapart.step = FALSE,
  betapart.para = FALSE,
  betapart.para.opt = betapart::beta.para.control()
)
}
\arguments{
\item{sp_faxes_coord}{a \bold{matrix} with coordinates of species (rows) on
functional axes (columns). Species coordinates have been retrieved thanks to
\code{\link{tr.cont.fspace}} or \code{\link{quality.fspaces}}.}

\item{asb_sp_occ}{a \bold{matrix} with presence/absence (coded as 0/1)
of species (columns) in a set of assemblages (rows).}

\item{check.input}{a \bold{logical value} defining whether inputs are checked before
computation of indices. Possible error messages will thus may be more
understandable for the user than R error messages. Default: check.input =
TRUE.}

\item{beta.family}{a \bold{character string} for the type of beta-diversity index to
use, 'Jaccard' (default) and/or 'Sorensen'}

\item{store_details}{a \bold{logical value} indicating whether the user wants to
store details. Details are used in the graphical function
\code{beta.multidim.plot} and thus must be kept if the user want to have
graphical outputs for the computed indices.}

\item{betapart.step}{a \bold{logical value} indicating whether the computation
progress tracking file "step.fbc.txt" should be created. Setting it to FALSE
will speed up the function. Default is FALSE, and it is automatically turned
to FALSE when 'betapart.para' is TRUE.}

\item{betapart.para}{a \bold{logical value} indicating whether internal
parallelization should be used to compute pairwise dissimilarities. Default
is FALSE.}

\item{betapart.para.opt}{a \bold{list} with details about parallelization. Default
value means those parameters are set according to computer specifications.
'nc' is number of cores (default = 4), 'type' is a character string with
code of method used (default PSOCK), 'LB' is a boolean specifying whether
load-balancing is applied (default is TRUE) and 'size' is a numeric value
for number of tasks performed at each time (default is 1). See help of
\code{\link[betapart]{functional.betapart.core}} for more details.}
}
\value{
a list with: \itemize{
\item inputs a list with \item \emph{pairasb_fbd_indices} a dataframe with for all pairs of
assemblages (rows) names of the assemblages in the 2 first columns and
columns with beta functional diversity indices following the additive
decomposition of Villéger et al 2013 : 'F_diss': functional dissimilarity
index, 'F_turn': its turnover component, and 'F_nest': its
nestedness-resultant component, i.e. F_diss = F_turn + F_nest. Those indices
names are preceded by the abbreviation of the type of indices used, 'jac'
for decomposition of Jaccard-like functional dissimilarity and 'sor' for
Sorensen-like dissimilarity.
\item \emph{details_beta} list if \emph{store_details} is TRUE:
\emph{sp_faxes_coord} and \emph{asb_sp_occ} on which indices were computed (convenient for drawing
graphics) ; a \bold{asb_FRic_raw} vector with volume of
the convex hull shaping each assemblage ; a \bold{asb_FRic} vector with volume of
the convex hull shaping each assemblage ; a \bold{asb_vertices} a list of
vectors (1 per assemblage) with names of species being vertices of the
convex hull }
}
\description{
Computes a set of indices of pairwise functional beta-diversity (dissimilarity
and its turnover and nestedness-resultant components) based on overlap between
convex hulls in a multidimensional space. For details about indices formulas
see Villeger et al. 2013, Global Ecology and Biogeography (22:671-681).
}
\section{Notes}{
 \bold{All assemblages should have a number of species strictly
higher than the number of functional axes}. Computing intersection of convex
hulls in space of >5 dimensions is yet impossible with most computers.
This function uses R libraries 'betapart' (>=1.5.2) for indices computation
and 'dendextend' for formatting values in a dataframe.
}

\examples{
load(system.file("extdata", "sp_faxes_coord_fruits", package = "mFD"))
load(system.file("extdata", "asb_sp_w_fruits", package = "mFD"))
asb_sp_w <- as.matrix(asb_sp_w)
asb_sp_summ <- mFD::asb.sp.summary(asb_sp_w)
asb_sp_occ <- asb_sp_summ$asb_sp_occ
beta_fd_indices <- mFD::beta.fd.multidim(sp_faxes_coord[, 
 c("PC1", "PC2", "PC3", "PC4")], asb_sp_occ,
 check.input = TRUE,
 beta.family = c("Jaccard"),
 store_details = TRUE)

}
