% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_used_for_ind_comp.R
\name{dist.point}
\alias{dist.point}
\title{Compute distances of all points to a given point in the functional space}
\usage{
dist.point(sp_faxes_coord, ref_sp)
}
\arguments{
\item{sp_faxes_coord}{a matrix of species coordinates in a chosen functional
space. Species coordinates have been retrieved thanks to
\code{\link{tr.cont.fspace}} or \code{\link{quality.fspaces}}.}

\item{ref_sp}{a character string referring to the name of the reference
species}
}
\value{
a vector of species distances to the reference species
\code{dist_refsp}
}
\description{
This function computes the distances of all species to a reference species.
It is used in FSpe, FOri and FNND computation.
}
\author{
Camille Magneville and Sébastien Villéger
}
