% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_indiv_ind_computation.R
\name{fdis.computation}
\alias{fdis.computation}
\title{Compute functional dispersion}
\usage{
fdis.computation(
  asb_sp_relatw_k,
  sp_faxes_coord_k,
  fide_asb = NULL,
  k,
  check.input = check.input
)
}
\arguments{
\item{asb_sp_relatw_k}{a \bold{matrix} containing species relative weight (columns)
for a given assemblage.}

\item{sp_faxes_coord_k}{a \bold{matrix} of species coordinates present in a given
assemblage in a chosen functional space with only needed axes. Species
coordinates have been retrieved thanks to \code{fspace.conttr} or
\code{qual.funct.space} and filtered thanks to \code{sp.filter}.}

\item{fide_asb}{a \bold{matrix} containing functional identity values a given
assemblage along the dimensions (columns). Can be retrieved after
\code{fide.computation} function and is compute if NULL. Default: fide_asb
= NULL.}

\item{k}{a \bold{character string} referring to the assemblage studied.}

\item{check.input}{a \bold{logical value} allowing to test or not the inputs.
Possible error messages will thus may be more understandable for the user
than R error messages. Species coordinates matrix and species*weight
dataframe must not contain NA, their rownames must be filled and they must
have similar names values. Default: check.input = FALSE.}
}
\value{
a matrix containing functional dispersion for a given assemblage.
}
\description{
This function computes the weighted deviation to center of gravity of species
present in a given assemblage. It is the weighted mean distance to the
weighted centroid. Its calculation requires FIde computation that can be
achieved thanks to \code{fide.computation} function. FDis value can be scaled
by the maximum value possible given species pool (i.e. the most distant
species pair have half of total weight) to standardize values.
}
