% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multidim_layers.R
\name{fdis.plot}
\alias{fdis.plot}
\title{Plot FDis index}
\usage{
fdis.plot(
  ggplot_bg,
  asb_sp_coord2D,
  asb_sp_relatw,
  asb_centroid_coord2D,
  plot_sp = TRUE,
  shape_sp,
  color_sp,
  fill_sp,
  shape_centroid,
  size_centroid,
  color_centroid,
  fill_centroid,
  color_segment,
  width_segment,
  linetype_segment
)
}
\arguments{
\item{ggplot_bg}{a ggplot object of the plot background retrieved through
the \code{\link{background.plot}} function.}

\item{asb_sp_coord2D}{a list of matrix (ncol = 2) with coordinates of
species present in each assemblage for a given pair of functional axes.}

\item{asb_sp_relatw}{a list of vector gathering species relative weight in
each assemblage. It can be retrieved through the
\code{\link{alpha.fd.multidim}}. If several assemblages it should be a
vector with names as in \code{asb_sp_coord2D}.}

\item{asb_centroid_coord2D}{a list (with names as in asb_sp_coord2D) of
vectors with coordinates of the centroid of species for each assemblage
for a given pair of axes.}

\item{plot_sp}{a logical value indicating whether species of each assemblage
should be plotted or not. Default: \code{plot_sp = TRUE}}

\item{shape_sp}{a numeric value referring to the shape used to plot species
belonging to the studied assemblage. Default: \code{shape_sp = 16}
(filled circle). If several assemblages it should be a vector with
names as in \code{asb_sp_coord2D}.}

\item{color_sp}{a R color name or an hexadecimal code referring to the color
of species from the studied assemblage. Default: \code{color_sp = "#0072B2"}.
If several assemblages it should be a vector with names as in
\code{asb_sp_coord2D}.}

\item{fill_sp}{a R color name or an hexadecimal code referring to the colour
to fill species symbol (if \code{shape_sp} > 20). Default:
\code{fill_sp = '#0072B2'}. If several assemblages it should  be a vector with
names as in \code{asb_sp_coord2D}.}

\item{shape_centroid}{a numeric value referring to the shape used to
plot centroid of the studied assemblage. Default: \code{shape_centroid = 18}
(filled diamond). If several assemblages it should be a vector with
names as in \code{asb_sp_coord2D}.}

\item{size_centroid}{a numeric value referring to the size of species
centroid but not the plotted assemblage. Default: \code{size_sp = 1}. If
several assemblages it should be a vector with names as in \code{asb_sp_coord2D}.}

\item{color_centroid}{a R color name or an hexadecimal code referring to the
color of the species centroid from the studied assemblage. If several
assemblages it should be a vector with names as in \code{asb_sp_coord2D}.}

\item{fill_centroid}{a R color name or an hexadecimal code referring to
the colour to fill assemblage centroid symbol (if \code{shape_sp} > 20).
If several assemblages it should be a vector with names as in
\code{asb_sp_coord2D}.}

\item{color_segment}{a R color name or an hexadecimal code referring to the
color of the segment linking axes and centroid from the studied assemblage.
If several assemblages it should be a vector with names as in
\code{asb_sp_coord2D}.}

\item{width_segment}{a numeric value referring to the size of the segment
linking centroid and functional axes. Default: \code{width_segment = 1}.  If
several assemblages it should be a vector with names as in \code{asb_sp_coord2D}.}

\item{linetype_segment}{a character string referring to the linetype used to
draw the segment linking centroid and functional axes.
Default: \verb{linetype_segment = "dashed}.  If several assemblages it should be a
vector with names as in \code{asb_sp_coord2D}.}
}
\value{
a ggplot object showing FDis index for one or several assemblage(s)
and a given pair of functional axes.
}
\description{
This function plots FDis index for a given pair of functional axes and for
one or several assemblages. It adds segments between species relative weights
and assemblage cenntroid on the background plot.
}
\examples{
\dontrun{
}

}
\author{
Camille Magneville and Sébastien Villéger
}
