% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funct_dist_computation.R
\name{funct.dist}
\alias{funct.dist}
\title{Compute functional distance between species}
\usage{
funct.dist(sp_tr, tr_cat, dist_metric)
}
\arguments{
\item{sp_tr}{a dataframe of traits values (columns) for each species (rows)}

\item{tr_cat}{a dataframe containing three columns for each trait (rows):
\itemize{
\item \strong{trait_name}: names of all traits as in \code{sp_tr} data.frame
\item \strong{trait_type}: category codes for each trait as followed:
\emph{N} for Nominal traits (factor variable), \emph{O} for Ordinal traits
(ordered variable), \emph{C} for Circular traits (integer values), \emph{Q}
for quantitative traits (numeric values) that is allowed \strong{only} if
there are at least 2 species with the same value and \emph{F} for fuzzy
traits (described with several values defined with several column).
}
An option is to add a fourth column with a numeric vector of length n
(traits number) to specify a weight for each trait. The name of the fourth
column has to be 'weight'.}

\item{dist_metric}{if euclidean, use euclidean distance,
classical_gower: use classical gower distance \emph{(J. C. (1971))}
kgower: gower distance modified by Pavoine \emph{Pavoine, S. et al (2009)}
Based on the \code{\link[ade4]{dist.ktab}} function: "If kgower
is chosen user have several choice to scale quantitative variables A string
that can have three values: either \code{"scaledBYrange"} if the quantitative
variables must be scaled by their range, or \code{"scaledBYsd"} if they must be
scaled by their standard deviation, or \code{"noscale"} if they should not be
scaled. This last option can be useful if the values have already been
normalized by the known range of the whole population instead of the observed
range measured on the sample. If x contains data from various types, then the
option \code{"scaledBYsd"} is not suitable (a warning will appear if the option
selected with that condition)."}
}
\value{
a dist object containing distance between each pair of species
}
\description{
For a given combination of traits, this function returns the functional
distance matrix between species
}
\examples{
load(system.file("extdata", "sp_tr_fruits_df", package = "mFD"))
sp_tr <- sp_tr[, -c(6:8)]
load(system.file("extdata", "sp_tr_cat_fruits_df", package = "mFD"))
sp_tr_cat <- sp_tr_cat[-c(6:8), ]
mFD::funct.dist(sp_tr, sp_tr_cat, dist_metric = "classical_gower")

}
