% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFD-package.R
\docType{package}
\name{mFD-package}
\alias{mFD}
\alias{mFD-package}
\title{mFD: Compute and Illustrate the Multiple Facets of Functional Diversity}
\description{
A computation of functional traits-based distances between pairs
    of species for species gathered in assemblages allowing to build several
    functional spaces. The package allows to compute functional diversity
    indices assessing the distribution of species (and of their dominance) in a
    given functional space for each assemblage and the overlap between
    assemblages in a given functional space. Graphical outputs included. See
    the mFD website for more information, documentation and examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CmlMagneville/mFD}
  \item Report bugs at \url{https://github.com/CmlMagneville/mFD/issues}
}

}
\author{
\strong{Maintainer}: Camille Magneville \email{camille.magneville@gmail.com} (\href{https://orcid.org/0000-0003-0489-3822}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicolas Loiseau \email{nicolas.loiseau@umontpellier.fr} (\href{https://orcid.org/0000-0002-2469-1980}{ORCID})
  \item Camille Albouy \email{albouycamille@gmail.comr} (\href{https://orcid.org/0000-0003-1629-2389}{ORCID})
  \item Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})
  \item Thomas Claverie \email{tclaverie@gmail.com} (\href{https://orcid.org/0000-0002-6258-4991}{ORCID})
  \item Arthur Escalas \email{arthur.escalas@umontpellier.fr} (\href{https://orcid.org/0000-0002-6450-5716}{ORCID})
  \item Fabien Leprieur \email{fabien.leprieur@umontpellier.fr} (\href{https://orcid.org/0000-0001-6869-342X}{ORCID})
  \item Eva Maire \email{emg.maire@umontpellier.fr} (\href{https://orcid.org/0000-0002-1032-3394}{ORCID})
  \item David Mouillot \email{david.mouillot@gmail.com} (\href{https://orcid.org/0000-0003-0402-2605}{ORCID})
  \item Sebastien Villeger \email{sebastien.villeger@cnrs.fr} (\href{https://orcid.org/0000-0002-2362-7178}{ORCID})
}

}
\keyword{internal}
