% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFD_utility.R
\docType{package}
\name{MFD-package}
\alias{MFD}
\alias{MFD-package}
\title{MFD: Decision-guided multi-ancestry fine-mapping}
\description{
Utilities for decision-guided multi-ancestry fine-mapping using
SuSiE post-hoc and MESuSiE.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
