% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_used_for_ind_comp.R
\name{mst.computation}
\alias{mst.computation}
\title{Compute the minimum spanning tree (MST) linking species of a given assemblage}
\usage{
mst.computation(sp_faxes_coord_k)
}
\arguments{
\item{sp_faxes_coord_k}{a \strong{matrix} relating species coordinates for species
present in a given assemblage}
}
\value{
a dist object summarizing the MST for all species of a given
assemblage \code{mst_asb_k}
}
\description{
This function computes the MST linking species of a given assemblage and is
used to compute FEve index.
}
