% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_filtering.R
\name{sp.filter}
\alias{sp.filter}
\title{Retrieve information about species in a given assemblage}
\usage{
sp.filter(asb_nm, sp_faxes_coord, asb_sp_w)
}
\arguments{
\item{asb_nm}{a \strong{string} object referring to the name of a given
community.}

\item{sp_faxes_coord}{a \strong{matrix} of species coordinates in a chosen functional
space. Species coordinates have been retrieved thanks to
\code{\link{tr.cont.fspace}} or \code{\link{quality.fspaces}}.}

\item{asb_sp_w}{a \strong{matrix} linking weight of species (columns) and a set of
assemblages (rows).}
}
\value{
a vector containing names of species present in a given assemblage \code{sp_name_asb_k}, a
matrix containing coordinates of species present in a given assemblage \code{sp_faxes_coord_k}, a
matrix containing weight of species present in a given assemblage \code{asb_sp_w_k}, a matrix
containing relative weight of species present in a given assemblage \code{asb_sp_relatw_k}.
}
\description{
This function computes names of species present in an given assemblage, their
coordinates in the functional space and their weights. It is used in the
\code{alpha_FD_multidim} function to filter species and compute each
functional indices for each community.
}
\examples{
 # Load Species*Traits dataframe:
data("sp_tr_fruits", package = "mFD")
# Load Assemblages*Species dataframe:      
data("asb_sp_w_fruits", package = "mFD") 
Load Traits categories dataframe:
data("tr_cat_fruits", package = "mFD")  
# Compute functional distance 
sp_dist_fruits <- mFD::funct.dist(sp_tr = sp_tr_fruits,         
 tr_cat       = tr_cat_fruits,   
 dist_metric  = "kgower",         
 scaling      = "scaledBYrange",  
 stop_if_NA   = TRUE)
# Compute functional spaces quality to retrieve species coordinates matrix:
fspaces_quality_fruits <- mFD::quality.fspaces(sp_dist = sp_dist_fruits, 
 maxdim_pcoa         = 10,
 deviation_weighting = "absolute",
 fdist_scaling       = FALSE,
 fdendro             = "average")
# Retrieve species coordinates matrix:
sp_faxes_coord_fruits <- fspaces_quality_fruits$"details_fspaces"$"sp_pc_coord"
# Filter species of basket_1 assemblage:
sp.filter(asb_nm = "basket_1", sp_faxes_coord_fruits, asb_sp_w_fruits)

}
