% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_summaries.R
\name{sp.tr.summary}
\alias{sp.tr.summary}
\title{Summarize Species x Traits Data Frame}
\usage{
sp.tr.summary(tr_cat, sp_tr)
}
\arguments{
\item{tr_cat}{a data frame containing three columns for each trait (rows):
\itemize{
\item \strong{trait_name}: the name of all traits as in \code{sp_tr} data
frame;
\item \strong{trait_type}: the category code for each trait as followed:
\code{N} for Nominal traits (factor variable), \code{O} for Ordinal traits
(ordered variable), \code{C} for Circular traits (integer values), \code{Q}
for quantitative traits (numeric values), and \code{F} for
fuzzy traits (i.e. described with several values defined with several
column);
\item \strong{fuzzy_name}: name of fuzzy-coded trait to which 'sub-trait'
belongs (if trait is not fuzzy, ignored so could be trait name or NA).
}

An option is to add a fourth column with a numeric vector of length n
(traits number) to specify a weight for each trait.}

\item{sp_tr}{a data frame of traits values (columns) for each species (rows).
Note that species names \strong{must be} specified in the row names.}
}
\value{
If there is no fuzzy-coded trait, a three-elements list with:
\item{tr_summary_list}{a table summarizing for each trait the
number of species per modality for non-continuous trait and min, max,
mean, median, and quartiles for continuous traits.}
\item{tr_types}{a list containing traits type.}
\item{non_conttr_mod_list}{a list containing modalities for non-continuous
trait.}

If there is fuzzy-coded trait, a four-elements list with:
\item{tr_summary_non_fuzzy_list}{a table summarizing for each trait the
number of species per modality for non-continuous trait and min, max,
mean, median, and quartiles for continuous traits.}
\item{tr_summary_fuzzy_list}{\strong{MISSING DESCRIPTION}}
\item{tr_types}{a list containing traits type.}
\item{mod_list}{a list containing modalities for non-continuous trait.}
}
\description{
This function computes a summary data helping to choose the type of analysis
you can do with your data. For this function to work, there must be no NA in
your \code{sp_tr} data frame.
}
\examples{
load(system.file("extdata", "sp_tr_fruits_df", package = "mFD"))
load(system.file("extdata", "sp_tr_cat_fruits_df", package = "mFD"))

mFD::sp.tr.summary(tr_cat = sp_tr_cat, sp_tr = sp_tr)
}
\author{
Camille Magneville & Sebastien Villeger
}
