% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_tr_fruits.R
\docType{data}
\name{sp_tr_fruits}
\alias{sp_tr_fruits}
\title{Traits Values of Fruits Species}
\format{
A data frame with 25 rows (species) and the following columns
(traits):
\describe{
\item{Size}{an ordered factor describing the size of fruits species}
\item{Plant}{an unordered factor describing the type of plant}
\item{Climate}{an ordered factor describing the climate regions}
\item{Seed}{an ordered factor describing the type of seed}
\item{Sugar}{a numeric describing the quantity of sugar}
\item{Use.raw}{an integer (percentage) describing the proportion of a raw
use (fuzzy trait) of the fruit}
\item{Use.pastry}{an integer (percentage) describing the proportion of a
pastry use (fuzzy trait) of the fruit}
\item{Use.jam}{an integer (percentage) describing the proportion of a
jam use (fuzzy trait) of the fruit}
}
}
\usage{
sp_tr_fruits
}
\description{
This dataset represents the value of 6 traits for 15 fruits species.
\strong{Important:} Species names must be specified as the data frame row names
(not in an additional column).
}
\examples{
# Load Species x Traits Data Frame
data("sp_tr_fruits", package = "mFD")
sp_tr_fruits

# Load Traits Information
data("tr_cat_fruits", package = "mFD")
tr_cat_fruits

# Summarize Species x Traits Data
mFD::sp.tr.summary(tr_cat = tr_cat_fruits, sp_tr = sp_tr_fruits)
}
\seealso{
\code{tr_cat_fruits}, \code{asb_sp_w_fruits}
}
\keyword{datasets}
