% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_cat_fruits.R
\docType{data}
\name{tr_cat_fruits}
\alias{tr_cat_fruits}
\title{Fruits Traits Informations}
\format{
A data frame with 8 rows (traits) and the following three columns:
\describe{
\item{trait_name}{a character giving the trait name}
\item{trait_type}{a character giving the trait type (\strong{O} for Ordinal
trait, \strong{N} for Nominal trait, \strong{Q} for Quantitative trait, and \strong{F} for
Fuzzy-coded trait)}
\item{fuzzy_name}{a character giving the name of fuzzy-coded traits
(i.e. \code{Use}) to  which 'sub-traits' (i.e. \code{raw}, \code{pastry}, and \code{jam})
belongs}
}
}
\usage{
tr_cat_fruits
}
\description{
This dataset summarizes information about the 6 traits used in the
\code{sp_tr_fruits} dataset.
}
\note{
If your dataset does not contain fuzzy trait, the column \code{fuzzy_name}
can be ignored but the first two columns are mandatory.

Traits in this dataset correspond to columns (traits) of the \code{sp_tr_fruits}
dataset.
}
\examples{
# Load Traits Information
data("tr_cat_fruits", package = "mFD")
tr_cat_fruits

# Load Species x Traits Data Frame
data("sp_tr_fruits", package = "mFD")

# Summarize Species x Traits Data
mFD::sp.tr.summary(tr_cat = tr_cat_fruits, sp_tr = sp_tr_fruits)
}
\seealso{
\code{sp_tr_fruits}, \code{asb_sp_w_fruits}
}
\keyword{datasets}
